update: i type (int to size_t)

+++ src/parse_utils.c
@@ -1,6 +1,6 @@
 #include "parse.h"
 
-void	skip_space(char *line, int *i)
+void	skip_space(char *line, size_t *i)
 {
 	while (line[*i] != '\0' && ft_strchr(" \t", line[*i]) != NULL)
 		(*i)++;
